Add: Support holding multiple credentials
