+++ build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
     id("com.android.application").apply(false)
     id("com.android.library").apply(false)
     id("org.jetbrains.compose").apply(false)
-    id("at.asitplus.gradle.conventions") version "1.9.23+20240319"
+    id("at.asitplus.gradle.conventions") version "1.9.23+20240501"
 }
 
 repositories {
