+++ shared/src/commonMain/composeResources/values-en/strings.xml
@@ -116,6 +116,7 @@
     <string name="heading_label_data_protection">Privacy policy</string>
     <string name="heading_label_terms_of_use_and_data_protection">Terms of use &amp; Privacy policy</string>
     <string name="heading_label_load_data_screen">Load data</string>
+    <string name="heading_label_add_credential_screen">Load data</string>
     <string name="heading_label_refresh_data_screen">Load data</string>
     <string name="heading_label_authenticate_at_device_screen">Authenticate at\ncounter or mashine</string>
 
@@ -127,6 +128,7 @@
     <string name="biometric_authentication_prompt_to_load_data_subtitle">Load Data</string>
 
     <string name="content_description_navigate_back">Back</string>
+    <string name="content_description_delete_credential">Delete</string>
     <string name="content_description_portrait">Portrait</string>
     <string name="content_description_refresh_credentials">Load Data</string>
     <string name="content_description_add_credential">Load more data</string>
@@ -168,6 +170,7 @@
     <string name="info_text_redirection_to_id_austria_for_credential_provisioning">To load your data you will be redirected to ID Austria.</string>
     <string name="info_text_notice_development_provisioning_using_qr_code_credentials">For development purposes, there is also the option to load data via a QR code.</string>
     <string name="info_text_enthusiastic_welcome_end">Load your data and off you go!</string>
+    <string name="info_text_submission_preview_disabled">Preview of data to be submitted is currently not available.</string>
 
     <string name="section_heading_authenticate_at_device_title">Authenticate</string>
     <string name="section_heading_authenticate_at_device_subtitle">at counters or mashines</string>
@@ -192,6 +195,7 @@
     <string name="section_heading_identity_data_icon_text">ID</string>
     <string name="section_heading_residence_data">Residence</string>
     <string name="section_heading_residence_data_icon_text">R</string>
+    <string name="section_heading_load_data_selection">Available data</string>
     <string name="section_heading_missing_data">Load new data</string>
     <string name="section_heading_available_data">Refresh existing data</string>
