+++ shared/src/commonMain/kotlin/ui/composables/ScreenHeading.kt
@@ -1,0 +1,13 @@
+package ui.composables
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ScreenHeading(title: String) {
+    Text(
+        text = title,
+        style = MaterialTheme.typography.headlineLarge,
+    )
+}
