+++ shared/src/commonMain/kotlin/ui/views/AuthenticationConsentView.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -15,8 +14,6 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.BottomAppBarDefaults
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
@@ -37,26 +34,18 @@ import composewalletapp.shared.generated.resources.biometric_authentication_prom
 import composewalletapp.shared.generated.resources.biometric_authentication_prompt_for_data_transmission_consent_title
 import composewalletapp.shared.generated.resources.heading_label_authenticate_at_device_screen
 import composewalletapp.shared.generated.resources.heading_label_navigate_back
-import composewalletapp.shared.generated.resources.prompt_ask_load_missing_data
+import composewalletapp.shared.generated.resources.info_text_submission_preview_disabled
 import composewalletapp.shared.generated.resources.prompt_send_above_data
-import composewalletapp.shared.generated.resources.prompt_send_all_data
 import composewalletapp.shared.generated.resources.section_heading_data_recipient
-import composewalletapp.shared.generated.resources.section_heading_requested_data
-import composewalletapp.shared.generated.resources.warning_requested_data_not_available_content
-import composewalletapp.shared.generated.resources.warning_requested_data_not_available_heading
 import org.jetbrains.compose.resources.ExperimentalResourceApi
 import org.jetbrains.compose.resources.stringResource
-import ui.composables.AttributeAvailability
 import ui.composables.BiometryPrompt
 import ui.composables.BiometryPromptDismissResult
 import ui.composables.BiometryPromptSuccessResult
-import ui.composables.DataCategoryDisplaySection
 import ui.composables.DataDisplaySection
-import ui.composables.PersonalDataCategory
 import ui.composables.buttons.CancelButton
 import ui.composables.buttons.ConsentButton
 import ui.composables.buttons.NavigateUpButton
-import ui.composables.buttons.ReloadDataButton
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalResourceApi::class)
 @Composable
@@ -64,9 +53,7 @@ fun AuthenticationConsentView(
     spName: String,
     spLocation: String,
     spImage: ImageBitmap?,
-    requestedAttributes: List<Pair<PersonalDataCategory, List<AttributeAvailability>>>,
     navigateUp: () -> Unit,
-    loadMissingData: () -> Unit,
     consentToDataTransmission: () -> Unit,
     cancelAuthentication: () -> Unit,
     showBiometry: Boolean,
@@ -93,33 +80,13 @@ fun AuthenticationConsentView(
                     BottomAppBarDefaults.ContainerElevation
                 )
             ) {
-                val hasMissingAttributes = requestedAttributes.any {
-                    it.second.any {
-                        it.isAvailable == false
-                    }
-                }
-
                 Column(
                     horizontalAlignment = Alignment.CenterHorizontally,
                     verticalArrangement = Arrangement.SpaceEvenly,
                     modifier = Modifier.fillMaxWidth().padding(top = 16.dp),
                 ) {
-                    val bottomBarQuestion =
-                        if (hasMissingAttributes) {
-                            stringResource(Res.string.prompt_ask_load_missing_data)
-                        } else if (requestedAttributes.isNotEmpty()) {
-                            stringResource(Res.string.prompt_send_above_data)
-                        } else {
-                            stringResource(Res.string.prompt_send_all_data)
-                        }
-
-                    val bottomBarContinueButton: @Composable RowScope.() -> Unit = {
-                        ReloadDataButton(loadMissingData)
-                        ConsentButton(consentToDataTransmission)
-                    }
-
                     Text(
-                        text = bottomBarQuestion,
+                        text = stringResource(Res.string.prompt_send_above_data),
                         style = MaterialTheme.typography.bodyMedium,
                     )
                     Row(
@@ -128,8 +95,6 @@ fun AuthenticationConsentView(
                         horizontalArrangement = Arrangement.Center,
                     ) {
                         CancelButton(cancelAuthentication)
-                        Spacer(modifier = Modifier.width(16.dp))
-                        ReloadDataButton(loadMissingData)
                         Spacer(modifier = Modifier.width(16.dp))
                         ConsentButton(consentToDataTransmission)
                     }
@@ -159,36 +124,6 @@ fun AuthenticationConsentView(
                 Column(
                     modifier = Modifier.fillMaxSize().verticalScroll(state = rememberScrollState()),
                 ) {
-                    val hasMissingAttributes = requestedAttributes.any {
-                        it.second.any {
-                            it.isAvailable == false
-                        }
-                    }
-                    if (hasMissingAttributes) {
-                        ElevatedCard(
-                            colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.errorContainer),
-                            modifier = paddingModifier,
-                        ) {
-                            Text(
-                                text = stringResource(Res.string.warning_requested_data_not_available_heading),
-                                style = MaterialTheme.typography.titleMedium,
-                                modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-                            )
-                            Column(
-                                modifier = Modifier.padding(
-                                    top = 0.dp,
-                                    end = 16.dp,
-                                    bottom = 8.dp,
-                                    start = 16.dp
-                                )
-                            ) {
-                                Text(
-                                    text = stringResource(Res.string.warning_requested_data_not_available_content),
-                                    style = MaterialTheme.typography.bodyLarge,
-                                )
-                            }
-                        }
-                    }
                     if (spImage != null) {
                         Box(Modifier.fillMaxWidth(), contentAlignment = Alignment.Center) {
                             Image(
@@ -207,13 +142,11 @@ fun AuthenticationConsentView(
                         ).toList(),
                         modifier = paddingModifier,
                     )
-                    if (requestedAttributes.isNotEmpty()) {
-                        DataCategoryDisplaySection(
-                            title = stringResource(Res.string.section_heading_requested_data),
-                            attributes = requestedAttributes,
-                            modifier = paddingModifier,
-                        )
-                    }
+
+                    Text(
+                        stringResource(Res.string.info_text_submission_preview_disabled),
+                        modifier = Modifier.weight(1.0f, true)
+                    )
                 }
             }
         }
